#189 Split gtfs_utils to core_calculations and partridge_helper
