+++ gtfs/gtfs_utils/gtfs_utils/gtfs_stats.py
@@ -13,8 +13,8 @@ from typing import List, Dict
 from tqdm import tqdm
 from partridge import feed as ptg_feed
 from .files import get_dates_without_output
-from .gtfs_utils import write_filtered_feed_by_date, get_partridge_feed_by_date, get_zones_df, \
-    compute_route_stats, compute_trip_stats
+from .core_calculations import get_zones_df, compute_route_stats, compute_trip_stats
+from .partridge_helper import write_filtered_feed_by_date, get_partridge_feed_by_date
 from .environment import init_conf
 from .s3 import get_latest_file, fetch_remote_file
 from .logging_config import configure_logger
